The Goal: I want to add a special administrative privilege to my website that allows only specific owners to delete users.
The Condition:

There are two site owners: Haroon Hatem (haroonhatem34@gmail.com) and Hamzah Miser (hamzahmisr@gmail.com).

When one of these owners logs in, they should gain a special permission that allows them to delete any registered user (whether a graduate or a company).

Their identity must be verified (by their email address) after they log in.

The Requested Solution:
I would like to modify the current code to check the email address of the logged-in user. If the email matches one of the two owners' emails, an additional option or link should be displayed in the navigation bar (navbar.php) that grants them access to an admin panel where they can delete users.  After adding permissions to the owners of the site. The names or emails of the site's owners should not appear in the deletion menu